+++ install_piRa1n-web.sh
@@ -7,5 +7,5 @@ sudo systemctl restart apache2.service
 cd /home/pi/
 git clone https://github.com/raspberryenvoie/piRa1n-web.git
 cd piRa1n-web/
-sudo cp index.php options.php shutdown.php style.css stylesheet.css update.php update_status.php exit_recovery_mode.php /var/www/html/
-echo -e "\n# piRa1n-web\nwww-data ALL=(ALL) NOPASSWD: /home/pi/piRa1n/config.sh\nwww-data ALL=(ALL) NOPASSWD: /home/pi/piRa1n/shutdown.sh\nwww-data ALL=(ALL) NOPASSWD: /home/pi/piRa1n-web/update.sh\nwww-data ALL=(ALL) NOPASSWD: /home/pi/piRa1n/exit_recovery_mode.sh\n# End of piRa1n-web" | sudo EDITOR='tee -a' visudo
+sudo cp index.php options.php shutdown.php style.css stylesheet.css update.php update_status.php exit_recovery_mode.php odysseyra1n.php /var/www/html/
+echo -e "\n# piRa1n-web\nwww-data ALL=(ALL) NOPASSWD: /home/pi/piRa1n/config.sh\nwww-data ALL=(ALL) NOPASSWD: /home/pi/piRa1n/shutdown.sh\nwww-data ALL=(ALL) NOPASSWD: /home/pi/piRa1n-web/update.sh\nwww-data ALL=(ALL) NOPASSWD: /home/pi/piRa1n/exit_recovery_mode.sh\nwww-data ALL=(ALL) NOPASSWD: /home/pi/piRa1n/odysseyra1n.sh\nwww-data ALL=(ALL) NOPASSWD: /home/pi/piRa1n-web/odysseyra1n_done.sh\n# End of piRa1n-web" | sudo EDITOR='tee -a' visudo