+++ odysseyra1n_done.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+systemctl start piRa1n.service