+++ shutdown.php
@@ -13,10 +13,10 @@
     </p>
     <div class="output">
       <?php
-      if(isset($_POST['shutdownSubmit'])){
+      if (isset($_POST['shutdownSubmit'])){
         shell_exec("nohup sudo /home/pi/piRa1n/shutdown.sh > /dev/null 2>&1 &");
-        echo "Your Pi has been shut down!";
       }
+      Your Pi has been shut down!
       ?>
     </div>
   </body>