+++ update.php
@@ -15,7 +15,7 @@
        <strong>⚠️ Don't shutdown the Pi until the updates are completed!</strong><br>
        You can find the status of updates <a href="update_status.php" class="updateStatus">here</a>.
       <?php
-      if(isset($_POST['updateSubmit'])){
+      if (isset($_POST['updateSubmit'])){
         shell_exec("nohup sudo /home/pi/piRa1n-web/update.sh > /dev/null 2>&1 &");
       }
       ?>